refactor(api): split routes.js into domain modules + shared gateway core
Behavior-preserving restructure — no endpoint or response changes
(verified: identical 113 registered routes vs main).

- Split the ~1.8k-line api/routes.js into per-domain modules under
  api/routes/ (status, caps, keys, recommendations, evals, analytics,
  governance, …), with routes.js now a thin re-export.
- Extract a shared gateway/core.js (resolveRoute / invokeWithFallback /
  getAppConfig / setGatewayHeaders) so the native and OpenAI-compat
  entry points cannot drift on routing policy.
- openaiCompat.js now imports the shared core instead of defining its
  own header helper.
